Fix the ERROR methods for string and symbol.
Attempting to call error with a string or symbol results in an error
stating that there's no defined method matching the
parameters. Removing the :AROUND qualifiers from the string and symbol
methods fixes this.

Calling ERROR with NIL as an argument will cause to an infinite
recursion, since the :BEFORE method specializing on CONDITION will
signal a TYPE-ERROR if arguments exist, and that will result in
another call to ERROR with a NIL argument, etc. Calling with no
arguments fixes this.